+++ back/engines/commercial/insights/spec/policies/insights/view_policy_spec.rb
@@ -1,0 +1,61 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+RSpec.describe Insights::ViewPolicy, type: :policy do
+  subject { described_class.new(user, view) }
+
+  let!(:all_views) { create_list(:view, 2) }
+  let(:view) { all_views.first }
+  let(:scope) { described_class::Scope.new(user, Insights::View) }
+
+  context 'when user is admin' do
+    let(:user) { build(:admin) }
+
+    it { is_expected.to permit(:show) }
+    it { is_expected.to permit(:create) }
+    it { is_expected.to permit(:destroy) }
+    it { is_expected.to permit(:update) }
+    it { expect(scope.resolve.count).to eq(2) }
+  end
+
+  context 'when user moderates the project associated with the view' do
+    let(:user) { build(:project_moderator, projects: [view.scope]) }
+
+    it { is_expected.to permit(:show) }
+    it { is_expected.to permit(:create) }
+    it { is_expected.to permit(:destroy) }
+    it { is_expected.to permit(:update) }
+    it { expect(scope.resolve.count).to eq(1) }
+  end
+
+  context 'when user is a visitor' do
+    let(:user) { nil }
+
+    it { is_expected.not_to permit(:show) }
+    it { is_expected.not_to permit(:create) }
+    it { is_expected.not_to permit(:destroy) }
+    it { is_expected.not_to permit(:update) }
+    it { expect { scope.resolve.count }.to raise_error(Pundit::NotAuthorizedError) }
+  end
+
+  context 'when user is a regular user' do
+    let(:user) { build(:user) }
+
+    it { is_expected.not_to permit(:show) }
+    it { is_expected.not_to permit(:create) }
+    it { is_expected.not_to permit(:destroy) }
+    it { is_expected.not_to permit(:update) }
+    it { expect { scope.resolve.count }.to raise_error(Pundit::NotAuthorizedError) }
+  end
+
+  context 'when user is moderator of another project' do
+    let(:user) { build(:project_moderator) }
+
+    it { is_expected.not_to permit(:show) }
+    it { is_expected.not_to permit(:create) }
+    it { is_expected.not_to permit(:destroy) }
+    it { is_expected.not_to permit(:update) }
+    it { expect(scope.resolve.count).to eq(0) }
+  end
+end